requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-socketio>=5.10.0
python-engineio>=4.8.0
websockets>=11.0
pydantic>=2.0.0
pyyaml>=6.0
loguru>=0.7.0
zai-sdk>=0.2.2
